// src/services/paciente.service.d.ts
export declare const crearPaciente: (datos: {
    nombre: string;
    apellido: string;
    email: string;
    password: string;
    fechaNacimiento: Date;
    documento: number;
    numero: number;
}) => Promise<{
    id: unknown;
    nombre: string;
    apellido: string;
    email: string;
    fechaNacimiento: Date;
    documento: number;
    numero: number;
}>;
//# sourceMappingURL=paciente.service.d.ts.map